se realizan modificaciones para los metodos POST en Noticia y Producto

+++ flora_api/DTO/Producto/DTOProducto.cs
@@ -16,6 +16,6 @@ namespace api_flora.DTO.Producto
 
         public string Imagen { get; set; } = "";
 
-        public int IdCategoria {  get; set; } = 0; 
+        public long IdCategoria {  get; set; } = 0; 
     }
 }
